added new diffum way
