+++ package.json
@@ -3,7 +3,7 @@
   "version": "1.0.0",
   "main": "index.js",
   "scripts": {
-    "start": "cross-env APP_ENV=prod NODE_ENV=production node ./newWay.js"
+    "start": "cross-env APP_ENV=prod NODE_ENV=production node ./index.js"
   },
   "author": "",
   "license": "ISC",
