+++ update/utils.js
@@ -1,3 +1,26 @@
+
+
+function localDiffum(image_dataArr,diffum_color,cant_pix_xday,imageInfo){
+    let untouched_pix=get_untouchedPix(image_dataArr,diffum_color,imageInfo.channels);
+    let new_image_dataArr=image_dataArr
+
+    if (untouched_pix.length<cant_pix_xday){
+        cant_pix_xday=untouched_pix.length;
+    }
+
+    for (let i=0;i<cant_pix_xday;i++){
+        let rand_arrPos=get_randNum(0,untouched_pix.length-1);
+        let pixel_coords=untouched_pix[rand_arrPos];
+        changePixel(pixel_coords[0],pixel_coords[1],diffum_color,image_dataArr,imageInfo);
+        delete_arrElem(rand_arrPos,untouched_pix);
+    }
+    
+    new_image_dataArr=image_dataArr;
+    return new_image_dataArr
+}
+
+
+
 function get_randNum(minimo,maximo){
     minimo = Math.ceil(minimo);
     maximo = Math.floor(maximo);
@@ -5,12 +28,23 @@ function get_randNum(minimo,maximo){
     return Math.floor(Math.random() * (maximo - minimo + 1)) + minimo;
 }
 
-function get_pixelCoords(arr){
-    let [x,y]=arr.slice(0,2);
+function get_untouchedPix(image_dataArr,diffum_color,channels){
+    let untouched_pix=[]
 
-    return [x,y];
+    for (let i=0;i<image_dataArr.length;i+=channels){
+        pixelColor=image_dataArr.slice(i,i+channels);
+        if (pixelColor[0]===diffum_color[0] && pixelColor[1]===diffum_color[1] && pixelColor[2]===diffum_color[2]){
+            let a=1
+        }
+        else{
+            let x=Math.floor(i/channels % 100); // Assuming width is 100
+            let y=Math.floor(i/channels / 100);
+            untouched_pix.push([x,y]);
+        }
+
+    }
+    return untouched_pix
 }
-
 
 function changePixel(x,y,new_color,imgArr,imgInfo){
     // Calcular el índice del píxel en el buffer
@@ -27,4 +61,4 @@ function delete_arrElem(index,arr){
 }
 
 
-module.exports={get_randNum,get_pixelCoords,changePixel,delete_arrElem};
+module.exports={localDiffum};
